+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/openid/PreRegisteredClientTest.kt
@@ -1,14 +1,15 @@
 package at.asitplus.wallet.lib.openid
 
-import at.asitplus.openid.*
-import at.asitplus.signum.indispensable.josef.*
+import at.asitplus.openid.AuthenticationRequestParameters
+import at.asitplus.openid.AuthenticationResponseParameters
+import at.asitplus.openid.OpenIdConstants
+import at.asitplus.openid.RequestParametersFrom
+import at.asitplus.signum.indispensable.josef.JwsSigned
 import at.asitplus.wallet.lib.agent.*
 import at.asitplus.wallet.lib.data.AtomicAttribute2023
 import at.asitplus.wallet.lib.data.ConstantIndex
 import at.asitplus.wallet.lib.data.vckJsonSerializer
-import at.asitplus.wallet.lib.jws.DefaultJwsService
 import at.asitplus.wallet.lib.jws.DefaultVerifierJwsService
-import at.asitplus.wallet.lib.oidc.RequestObjectJwsVerifier
 import at.asitplus.wallet.lib.oidvci.*
 import com.benasher44.uuid.uuid4
 import io.kotest.assertions.throwables.shouldThrow
@@ -23,23 +24,23 @@ import io.kotest.matchers.string.shouldNotContain
 import io.kotest.matchers.string.shouldStartWith
 import io.kotest.matchers.types.shouldBeInstanceOf
 import io.ktor.http.*
-import kotlinx.datetime.Clock
-import kotlin.time.Duration.Companion.seconds
 
-class OpenId4VpProtocolTest : FreeSpec({
+class PreRegisteredClientTest : FreeSpec({
 
     lateinit var clientId: String
     lateinit var walletUrl: String
+    lateinit var redirectUrl: String
     lateinit var holderKeyMaterial: KeyMaterial
     lateinit var verifierKeyMaterial: KeyMaterial
     lateinit var holderAgent: Holder
     lateinit var holderOid4vp: OpenId4VpHolder
     lateinit var verifierOid4vp: OpenId4VpVerifier
-
+    lateinit var defaultRequestOptions: RequestOptions
     beforeEach {
         holderKeyMaterial = EphemeralKeyWithoutCert()
         verifierKeyMaterial = EphemeralKeyWithoutCert()
-        clientId = "https://example.com/rp/${uuid4()}"
+        clientId = "PRE-REGISTERED-CLIENT"
+        redirectUrl = "https://example.com/rp/${uuid4()}"
         walletUrl = "https://example.com/wallet/${uuid4()}"
         holderAgent = HolderAgent(holderKeyMaterial)
 
@@ -58,13 +59,22 @@ class OpenId4VpProtocolTest : FreeSpec({
         )
         verifierOid4vp = OpenId4VpVerifier(
             keyMaterial = verifierKeyMaterial,
-            clientIdScheme = ClientIdScheme.RedirectUri(clientId),
+            clientIdScheme = ClientIdScheme.PreRegistered(clientId, redirectUrl),
+        )
+        defaultRequestOptions = RequestOptions(
+            credentials = setOf(
+                RequestOptionsCredential(ConstantIndex.AtomicAttribute2023)
+            ),
         )
     }
 
     "test with Fragment" {
         val authnRequest = verifierOid4vp.createAuthnRequest(
-            defaultRequestOptions, OpenId4VpVerifier.CreationOptions.Query(walletUrl)
+            RequestOptions(
+                credentials = setOf(RequestOptionsCredential(ConstantIndex.AtomicAttribute2023)),
+                responseMode = OpenIdConstants.ResponseMode.Fragment,
+            ),
+            OpenId4VpVerifier.CreationOptions.Query(walletUrl)
         ).getOrThrow().url
 
         val authnResponse = holderOid4vp.createAuthnResponse(authnRequest).getOrThrow()
@@ -72,19 +82,49 @@ class OpenId4VpProtocolTest : FreeSpec({
 
         authnResponse.url.shouldNotContain("?")
         authnResponse.url.shouldContain("#")
-        authnResponse.url.shouldStartWith(clientId)
+        authnResponse.url.shouldStartWith(redirectUrl)
 
         val result = verifierOid4vp.validateAuthnResponse(authnResponse.url)
             .shouldBeInstanceOf<AuthnResponseResult.Success>()
         result.vp.verifiableCredentials.shouldNotBeEmpty()
 
-        verifySecondProtocolRun(verifierOid4vp, walletUrl, holderOid4vp)
+        verifierOid4vp.createAuthnRequest(
+            defaultRequestOptions, OpenId4VpVerifier.CreationOptions.Query(walletUrl)
+        ).getOrThrow().url.let { newAuthnRequestUrl ->
+            verifySecondProtocolRun(
+                verifierOid4vp, newAuthnRequestUrl, holderOid4vp
+            )
+        }
+    }
+
+    "test with Query" {
+        val expectedState = uuid4().toString()
+        val authnRequest = verifierOid4vp.createAuthnRequest(
+            RequestOptions(
+                credentials = setOf(RequestOptionsCredential(ConstantIndex.AtomicAttribute2023)),
+                responseMode = OpenIdConstants.ResponseMode.Query,
+                state = expectedState,
+            ),
+            OpenId4VpVerifier.CreationOptions.Query(walletUrl)
+        ).getOrThrow().url
+
+        val authnResponse = holderOid4vp.createAuthnResponse(authnRequest).getOrThrow()
+            .shouldBeInstanceOf<AuthenticationResponseResult.Redirect>()
+
+        authnResponse.url.shouldContain("?")
+        authnResponse.url.shouldNotContain("#")
+        authnResponse.url.shouldStartWith(redirectUrl)
+
+        val result = verifierOid4vp.validateAuthnResponse(authnResponse.url)
+            .shouldBeInstanceOf<AuthnResponseResult.Success>()
+        result.vp.verifiableCredentials.shouldNotBeEmpty()
+        result.state.shouldBe(expectedState)
     }
 
     "wrong client nonce in id_token should lead to error" {
         verifierOid4vp = OpenId4VpVerifier(
             keyMaterial = verifierKeyMaterial,
-            clientIdScheme = ClientIdScheme.RedirectUri(clientId),
+            clientIdScheme = ClientIdScheme.PreRegistered(clientId, redirectUrl),
             nonceService = object : NonceService {
                 override suspend fun provideNonce() = uuid4().toString()
                 override suspend fun verifyNonce(it: String) = false
@@ -93,7 +133,7 @@ class OpenId4VpProtocolTest : FreeSpec({
         )
         val requestOptions = RequestOptions(
             credentials = setOf(RequestOptionsCredential(ConstantIndex.AtomicAttribute2023)),
-            responseType = OpenIdConstants.ID_TOKEN
+            responseType = OpenIdConstants.ID_TOKEN,
         )
         val authnRequest = verifierOid4vp.createAuthnRequest(
             requestOptions, OpenId4VpVerifier.CreationOptions.Query(walletUrl)
@@ -110,7 +150,7 @@ class OpenId4VpProtocolTest : FreeSpec({
     "wrong client nonce in vp_token should lead to error" {
         verifierOid4vp = OpenId4VpVerifier(
             keyMaterial = verifierKeyMaterial,
-            clientIdScheme = ClientIdScheme.RedirectUri(clientId),
+            clientIdScheme = ClientIdScheme.PreRegistered(clientId, redirectUrl),
             stateToAuthnRequestStore = object : MapStore<String, AuthenticationRequestParameters> {
                 override suspend fun put(key: String, value: AuthenticationRequestParameters) {}
                 override suspend fun get(key: String): AuthenticationRequestParameters? = null
@@ -156,14 +196,14 @@ class OpenId4VpProtocolTest : FreeSpec({
             RequestOptions(
                 credentials = setOf(RequestOptionsCredential(ConstantIndex.AtomicAttribute2023)),
                 responseMode = OpenIdConstants.ResponseMode.DirectPost,
-                responseUrl = clientId,
+                responseUrl = redirectUrl
             ),
             OpenId4VpVerifier.CreationOptions.Query(walletUrl)
         ).getOrThrow().url
 
         val authnResponse = holderOid4vp.createAuthnResponse(authnRequest).getOrThrow()
             .shouldBeInstanceOf<AuthenticationResponseResult.Post>()
-        authnResponse.url.shouldBe(clientId)
+        authnResponse.url.shouldBe(redirectUrl)
 
         val result = verifierOid4vp.validateAuthnResponse(authnResponse.params.formUrlEncode())
             .shouldBeInstanceOf<AuthnResponseResult.Success>()
@@ -175,14 +215,14 @@ class OpenId4VpProtocolTest : FreeSpec({
             RequestOptions(
                 credentials = setOf(RequestOptionsCredential(ConstantIndex.AtomicAttribute2023)),
                 responseMode = OpenIdConstants.ResponseMode.DirectPostJwt,
-                responseUrl = clientId,
+                responseUrl = redirectUrl
             ),
             OpenId4VpVerifier.CreationOptions.Query(walletUrl)
         ).getOrThrow().url
 
         val authnResponse = holderOid4vp.createAuthnResponse(authnRequest).getOrThrow()
             .shouldBeInstanceOf<AuthenticationResponseResult.Post>()
-        authnResponse.url.shouldBe(clientId)
+        authnResponse.url.shouldBe(redirectUrl)
         authnResponse.params.shouldHaveSize(2)
         val jarmResponse = authnResponse.params.entries.first { it.key == "response" }.value
         val jwsObject = JwsSigned.Companion.deserialize<AuthenticationResponseParameters>(
@@ -193,30 +233,6 @@ class OpenId4VpProtocolTest : FreeSpec({
         val result = verifierOid4vp.validateAuthnResponse(authnResponse.params.formUrlEncode())
             .shouldBeInstanceOf<AuthnResponseResult.Success>()
         result.vp.verifiableCredentials.shouldNotBeEmpty()
-    }
-
-    "test with Query" {
-        val expectedState = uuid4().toString()
-        val authnRequest = verifierOid4vp.createAuthnRequest(
-            RequestOptions(
-                credentials = setOf(RequestOptionsCredential(ConstantIndex.AtomicAttribute2023)),
-                responseMode = OpenIdConstants.ResponseMode.Query,
-                state = expectedState
-            ),
-            OpenId4VpVerifier.CreationOptions.Query(walletUrl)
-        ).getOrThrow().url
-
-        val authnResponse = holderOid4vp.createAuthnResponse(authnRequest).getOrThrow()
-            .shouldBeInstanceOf<AuthenticationResponseResult.Redirect>()
-
-        authnResponse.url.shouldContain("?")
-        authnResponse.url.shouldNotContain("#")
-        authnResponse.url.shouldStartWith(clientId)
-
-        val result = verifierOid4vp.validateAuthnResponse(authnResponse.url)
-            .shouldBeInstanceOf<AuthnResponseResult.Success>()
-        result.vp.verifiableCredentials.shouldNotBeEmpty()
-        result.state.shouldBe(expectedState)
     }
 
     "test with deserializing" {
@@ -268,52 +284,6 @@ class OpenId4VpProtocolTest : FreeSpec({
         result.vp.verifiableCredentials.shouldNotBeEmpty()
         result.vp.verifiableCredentials.forEach {
             it.vc.credentialSubject.shouldBeInstanceOf<AtomicAttribute2023>()
-        }
-    }
-
-    "test with request object and Attestation JWT" {
-        val sprsCryptoService = DefaultCryptoService(EphemeralKeyWithoutCert())
-        val attestationJwt = buildAttestationJwt(sprsCryptoService, clientId, verifierKeyMaterial)
-        verifierOid4vp = OpenId4VpVerifier(
-            keyMaterial = verifierKeyMaterial,
-            clientIdScheme = ClientIdScheme.VerifierAttestation(attestationJwt, clientId),
-        )
-        val authnRequestWithRequestObject = verifierOid4vp.createAuthnRequest(
-            requestOptionsAtomicAttribute(), OpenId4VpVerifier.CreationOptions.SignedRequestByValue(walletUrl)
-        ).getOrThrow().url
-
-        holderOid4vp = OpenId4VpHolder(
-            holder = holderAgent,
-            requestObjectJwsVerifier = attestationJwtVerifier(sprsCryptoService.keyMaterial.jsonWebKey)
-        )
-        val authnResponse = holderOid4vp.createAuthnResponse(authnRequestWithRequestObject).getOrThrow()
-            .shouldBeInstanceOf<AuthenticationResponseResult.Redirect>()
-
-        val result = verifierOid4vp.validateAuthnResponse(authnResponse.url)
-            .shouldBeInstanceOf<AuthnResponseResult.Success>()
-        result.vp.verifiableCredentials.shouldNotBeEmpty()
-        result.vp.verifiableCredentials.forEach {
-            it.vc.credentialSubject.shouldBeInstanceOf<AtomicAttribute2023>()
-        }
-    }
-    "test with request object and invalid Attestation JWT" {
-        val sprsCryptoService = DefaultCryptoService(EphemeralKeyWithoutCert())
-        val attestationJwt = buildAttestationJwt(sprsCryptoService, clientId, verifierKeyMaterial)
-
-        verifierOid4vp = OpenId4VpVerifier(
-            keyMaterial = verifierKeyMaterial,
-            clientIdScheme = ClientIdScheme.VerifierAttestation(attestationJwt, clientId)
-        )
-        val authnRequestWithRequestObject = verifierOid4vp.createAuthnRequest(
-            requestOptionsAtomicAttribute(), OpenId4VpVerifier.CreationOptions.SignedRequestByValue(walletUrl)
-        ).getOrThrow().url
-
-        holderOid4vp = OpenId4VpHolder(
-            holder = holderAgent,
-            requestObjectJwsVerifier = attestationJwtVerifier(EphemeralKeyWithoutCert().jsonWebKey)
-        )
-        shouldThrow<OAuth2Exception> {
-            holderOid4vp.createAuthnResponse(authnRequestWithRequestObject).getOrThrow()
         }
     }
 
@@ -395,64 +365,19 @@ class OpenId4VpProtocolTest : FreeSpec({
     }
 })
 
+
 private fun requestOptionsAtomicAttribute() = RequestOptions(
     credentials = setOf(
         RequestOptionsCredential(ConstantIndex.AtomicAttribute2023)
     ),
 )
 
-private suspend fun buildAttestationJwt(
-    sprsCryptoService: DefaultCryptoService,
-    clientId: String,
-    verifierKeyMaterial: KeyMaterial,
-): JwsSigned<JsonWebToken> = DefaultJwsService(sprsCryptoService).createSignedJws(
-    header = JwsHeader(
-        algorithm = sprsCryptoService.keyMaterial.signatureAlgorithm.toJwsAlgorithm().getOrThrow(),
-    ),
-    payload = JsonWebToken(
-        issuer = "sprs", // allows Wallet to determine the issuer's key
-        subject = clientId,
-        issuedAt = Clock.System.now(),
-        expiration = Clock.System.now().plus(10.seconds),
-        notBefore = Clock.System.now(),
-        confirmationClaim = ConfirmationClaim(jsonWebKey = verifierKeyMaterial.jsonWebKey),
-    ),
-    serializer = JsonWebToken.Companion.serializer(),
-).getOrThrow()
-
-private fun attestationJwtVerifier(trustedKey: JsonWebKey) =
-    object : RequestObjectJwsVerifier {
-        override fun invoke(jws: JwsSigned<RequestParameters>): Boolean {
-            val attestationJwt = jws.header.attestationJwt?.let {
-                JwsSigned.Companion.deserialize<JsonWebToken>(
-                    JsonWebToken.Companion.serializer(), it
-                ).getOrThrow()
-            }
-                ?: return false
-            val verifierJwsService = DefaultVerifierJwsService()
-            if (!verifierJwsService.verifyJws(attestationJwt, trustedKey))
-                return false
-            val verifierPublicKey = attestationJwt.payload.confirmationClaim?.jsonWebKey
-                ?: return false
-            return verifierJwsService.verifyJws(jws, verifierPublicKey)
-        }
-    }
-
 private suspend fun verifySecondProtocolRun(
     verifierOid4vp: OpenId4VpVerifier,
-    walletUrl: String,
+    authnRequestUrl: String,
     holderOid4vp: OpenId4VpHolder,
 ) {
-    val authnRequestUrl = verifierOid4vp.createAuthnRequest(
-        defaultRequestOptions, OpenId4VpVerifier.CreationOptions.Query(walletUrl)
-    ).getOrThrow().url
     val authnResponse = holderOid4vp.createAuthnResponse(authnRequestUrl)
     verifierOid4vp.validateAuthnResponse((authnResponse.getOrThrow() as AuthenticationResponseResult.Redirect).url)
         .shouldBeInstanceOf<AuthnResponseResult.Success>()
 }
-
-private val defaultRequestOptions = RequestOptions(
-    credentials = setOf(
-        RequestOptionsCredential(ConstantIndex.AtomicAttribute2023)
-    )
-)